requirements.txt
openai
python-dotenv #will be creating api key in env file
langchain_nvidia_ai_endpoints
langchain_community
faiss-cpu
streamlit
pypdf